Bazel client: introduce Path abstraction
The new Path class is a thin wrapper around
std::string. It provides a little extra type
safety at no memory cost.

Strings and string literals can implicitly be
converted to Path, but not the other way around.
This is important: to convert Path to string, you
must call a specific To...Path() method, so the
caller's intent with the result is obvious.

A Path abstraction benefits Windows. It enables:
- checking paths only once and saving on
  conversions (on Windows), and
- using wstring as the wrapped type (later), which
  is natural on Windows (because of the WCHAR
  Windows API), unlike the current string type

Change-Id: I60e9ce2ba894d0c0fcb28e1abc088ebd2aec83e0